refactor(web): merge checkin time stack into lib/format
checkin-time.ts duplicated lib/format.ts: parseServerUtcDate mirrored
toTimestamp and formatCheckinLogTime mirrored formatDateTime (identical
Intl field set). Now checkin is a thin forwarder:

- lib/format: export toTimestamp; add an internal formatWithParts core;
  formatDateTime gains an optional IANA timeZone param; new
  formatDateTimeMinute (minute precision) completes the vocabulary.
- checkin-time: parseServerUtcDateTime / formatCheckinLogTime /
  formatDateTimeMinuteLocal forward to lib; own parse+format stack deleted.
  datetime-local input helpers (check-in specific) stay.

Contract decision: invalid input now renders the shared em-dash (lib
contract) instead of the checkin-only raw-string passthrough. Blast radius
is minimal - the two live callers render server-generated created_at
timestamps - while keeping passthrough would have required loosening
formatDateTime for every lib consumer. Test updated with the rationale.

Verified with a before/after parity script: identical output for all valid
inputs across en-US/zh-CN x local/UTC/Asia/Shanghai; the only difference is
the documented invalid -> em-dash change.
