+++ web/src/features/checkin/lib/checkin-time.ts
@@ -1,108 +1,42 @@
-// metapi-go features/checkin/lib — UTC date parsing & formatting helpers.
+// metapi-go features/checkin/lib — check-in log time helpers.
 //
-// The backend stores `created_at` as a naive UTC string (e.g.
-// "2026-08-11 12:30:00") that may arrive without a timezone suffix. These
-// helpers normalise such values to real Dates before formatting them for the
-// UI, mirroring the legacy `pages/helpers/checkinLogTime` module which is
-// outside the `src/` alias scope and therefore cannot be imported here.
+// Parsing and formatting are delegated to the shared stack in `@/lib/format`
+// (`toTimestamp` + `formatDateTime` / `formatDateTimeMinute`) so the naive
+// UTC `created_at` values ("2026-08-11 12:30:00", no timezone suffix) are
+// interpreted exactly like every other timestamp in the admin console. This
+// module previously carried its own duplicate parse/format stack; it now only
+// keeps the check-in-specific `datetime-local` input helpers.
 
-function parseServerUtcDate(value: string | null | undefined): {
-  date: Date | null
-  raw: string
-} {
-  if (!value) return { date: null, raw: '' }
-
-  const raw = String(value).trim()
-  if (!raw) return { date: null, raw: '' }
-
-  // Epoch milliseconds / seconds shorthand.
-  if (/^\d{10,13}$/.test(raw)) {
-    const asNumber = Number(raw)
-    if (Number.isFinite(asNumber)) {
-      const epochMs = raw.length === 13 ? asNumber : asNumber * 1000
-      const date = new Date(epochMs)
-      if (!Number.isNaN(date.getTime())) return { date, raw }
-    }
-  }
-
-  let normalized = raw
-  if (!normalized.includes('T') && normalized.includes(' ')) {
-    normalized = normalized.replace(' ', 'T')
-  }
-  // Normalise bare offsets like +0800 → +08:00, or +08 → +08:00.
-  normalized = normalized.replace(/([+-]\d{2})(\d{2})$/, '$1:$2')
-  normalized = normalized.replace(/([+-]\d{2})$/, '$1:00')
-
-  const hasTimeZone = /[zZ]$|[+-]\d{2}:\d{2}$/.test(normalized)
-  if (!hasTimeZone) {
-    normalized = `${normalized}Z`
-  }
-
-  const date = new Date(normalized)
-  if (Number.isNaN(date.getTime())) return { date: null, raw }
-
-  return { date, raw }
-}
+import { formatDateTime, formatDateTimeMinute, toTimestamp } from '@/lib/format'
 
 export function parseServerUtcDateTime(
   value: string | null | undefined
 ): Date | null {
-  return parseServerUtcDate(value).date
+  const timestamp = toTimestamp(value)
+  return timestamp === null ? null : new Date(timestamp)
 }
 
-function formatWithParts(
-  value: string | null | undefined,
-  options: Intl.DateTimeFormatOptions,
-  locale: string,
-  timeZone?: string
-): string {
-  const { date, raw } = parseServerUtcDate(value)
-  if (!date) return raw || '—'
-  return new Intl.DateTimeFormat(locale, {
-    ...options,
-    ...(timeZone ? { timeZone } : {}),
-  }).format(date)
-}
-
+/**
+ * Format a check-in log timestamp with seconds. Thin forwarder to
+ * `formatDateTime`; unparseable input renders the shared "—" placeholder
+ * (the lib contract — the previous raw-string passthrough was retired so
+ * invalid data never leaks into the render tree).
+ */
 export function formatCheckinLogTime(
   value: string | null | undefined,
   locale: string,
   timeZone?: string
 ): string {
-  return formatWithParts(
-    value,
-    {
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit',
-      hour: '2-digit',
-      minute: '2-digit',
-      second: '2-digit',
-      hour12: false,
-    },
-    locale,
-    timeZone
-  )
+  return formatDateTime(value, locale, timeZone)
 }
 
+/** Minute-precision variant (seconds dropped). */
 export function formatDateTimeMinuteLocal(
   value: string | null | undefined,
   locale: string,
   timeZone?: string
 ): string {
-  return formatWithParts(
-    value,
-    {
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit',
-      hour: '2-digit',
-      minute: '2-digit',
-      hour12: false,
-    },
-    locale,
-    timeZone
-  )
+  return formatDateTimeMinute(value, locale, timeZone)
 }
 
 /**
